PyThumbnail/requirements.txt
pip==9.0.1
setuptools==28.8.0
Pillow
requests